# packages/api-spec/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vivliostyle Pub Sync API",
    "description": "Open API for Vivliostyle Pub project sync, files, attachments, and OAuth 2.1 authentication. Generated from the reference implementation (@v/api-server-reference); do not edit by hand.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to the server origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "postAuthRegister",
        "tags": ["auth"],
        "summary": "Register a new user (reference-server convenience).",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    }
                  },
                  "required": ["id", "username"]
                }
              }
            }
          },
          "409": {
            "description": "Username already taken",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 256
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        }
      }
    },
    "/oauth/authorize": {
      "post": {
        "operationId": "postOauthAuthorize",
        "tags": ["auth"],
        "summary": "Issue an authorization code (OAuth 2.1 + PKCE).",
        "description": "The reference server validates credentials inline in place of an interactive consent screen.",
        "responses": {
          "200": {
            "description": "Authorization code",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "redirectUri": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "redirectUri"]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientId": {
                    "type": "string"
                  },
                  "redirectUri": {
                    "type": "string",
                    "format": "uri"
                  },
                  "codeChallenge": {
                    "type": "string"
                  },
                  "codeChallengeMethod": {
                    "enum": ["S256", "plain"],
                    "type": "string",
                    "default": "S256"
                  },
                  "scope": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "clientId",
                  "redirectUri",
                  "codeChallenge",
                  "username",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/oauth/token": {
      "post": {
        "operationId": "postOauthToken",
        "tags": ["auth"],
        "summary": "Exchange an authorization code or refresh token for tokens.",
        "responses": {
          "200": {
            "description": "Tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "tokenType": {
                      "const": "Bearer"
                    },
                    "expiresIn": {
                      "type": "number"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "scope": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "accessToken",
                    "tokenType",
                    "expiresIn",
                    "refreshToken"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "grantType": {
                        "const": "authorization_code"
                      },
                      "code": {
                        "type": "string"
                      },
                      "codeVerifier": {
                        "type": "string"
                      },
                      "redirectUri": {
                        "type": "string",
                        "format": "uri"
                      },
                      "clientId": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "grantType",
                      "code",
                      "codeVerifier",
                      "redirectUri",
                      "clientId"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "grantType": {
                        "const": "refresh_token"
                      },
                      "refreshToken": {
                        "type": "string"
                      },
                      "clientId": {
                        "type": "string"
                      }
                    },
                    "required": ["grantType", "refreshToken", "clientId"]
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/oauth/refresh": {
      "post": {
        "operationId": "postOauthRefresh",
        "tags": ["auth"],
        "summary": "Refresh tokens (rotating refresh token).",
        "responses": {
          "200": {
            "description": "Tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "tokenType": {
                      "const": "Bearer"
                    },
                    "expiresIn": {
                      "type": "number"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "scope": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "accessToken",
                    "tokenType",
                    "expiresIn",
                    "refreshToken"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  },
                  "clientId": {
                    "type": "string"
                  }
                },
                "required": ["refreshToken", "clientId"]
              }
            }
          }
        }
      }
    },
    "/oauth/session": {
      "delete": {
        "operationId": "deleteOauthSession",
        "tags": ["auth"],
        "summary": "Revoke the caller's tokens.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/oauth/userinfo": {
      "get": {
        "operationId": "getOauthUserinfo",
        "tags": ["auth"],
        "summary": "Get the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    }
                  },
                  "required": ["id", "username"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
  }
}